revisian

+++ database/migrations/2024_10_28_173649_add_department_to_users.php
@@ -11,6 +11,7 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasColumn('users', 'department_id')) return;
         Schema::table('users', function (Blueprint $table) {
             $table->unsignedBigInteger('department_id');
             $table->foreign('department_id')->references('id')->on('departments')->onDelete('cascade');
@@ -22,8 +23,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            
-        });
+        Schema::table('users', function (Blueprint $table) {});
     }
 };
